feat: 문서 처리 기능 구현 및 코드 정리
- 문서 업로드 및 비동기 처리 파이프라인 구현
- Celery 태스크를 활용한 문서 처리 로직 개선
- 불필요한 코드 제거 및 리팩토링
