+++ app/tasks/document_tasks.py
@@ -4,11 +4,10 @@ from pathlib import Path
 from typing import List
 from uuid import UUID
 
-from app.main import app as fastapi_app
 from celery import Celery
 from sqlalchemy.ext.asyncio import AsyncSession
 
-from app.db.session import get_db_session
+from app.db.session import async_session_maker
 from app.crud.crud_document import document, document_chunk
 from app.models.db_models import DocumentProcessingStatus
 from app.schemas.document import DocumentChunkCreate, DocumentChunkUpdate
@@ -18,86 +17,80 @@ from langchain.text_splitter import RecursiveCharacterTextSplitter
 from langchain_openai import OpenAIEmbeddings
 from langchain_community.vectorstores import Chroma
 from app.core.config import settings
+from app.core.celery_utils import celery_app
 
 logger = logging.getLogger(__name__)
-
-celery_app: Celery = fastapi_app.state.celery_app
 
 
 @celery_app.task(name="process_document_task", bind=True)
 async def process_document(self, document_id: str, file_path: str):
     db: AsyncSession = None
     try:
-        async for session in get_db_session():
-            db = session
-            break
+        db = async_session_maker()
+        async with db as session:
+            doc_uuid = UUID(document_id)
+            db_document = await document.get(session, id=doc_uuid)
 
-        if not db:
-            raise Exception("Could not get database session")
+            if not db_document:
+                logger.error(f"Document with ID {document_id} not found.")
+                return
 
-        doc_uuid = UUID(document_id)
-        db_document = await document.get(db, id=doc_uuid)
-
-        if not db_document:
-            logger.error(f"Document with ID {document_id} not found.")
-            return
-
-        # Update document status to PROCESSING
-        await document.update(
-            db,
-            db_obj=db_document,
-            obj_in={
-                "processing_status": DocumentProcessingStatus.PROCESSING,
-                "error_message": None,
-            },
-        )
-        await db.commit()
-        await db.refresh(db_document)
-
-        # Load document
-        loader = None
-        file_extension = Path(file_path).suffix.lower()
-        if file_extension == ".pdf":
-            loader = PyPDFLoader(file_path)
-        elif file_extension == ".txt":
-            loader = TextLoader(file_path)
-        else:
-            raise ValueError(f"Unsupported file type: {file_extension}")
-
-        langchain_documents = loader.load()
-
-        # Split document into chunks
-        text_splitter = RecursiveCharacterTextSplitter(
-            chunk_size=1000, chunk_overlap=200
-        )
-        chunks = text_splitter.split_documents(langchain_documents)
-
-        # Create document chunks in DB and enqueue embedding tasks
-        for i, chunk in enumerate(chunks):
-            chunk_data = DocumentChunkCreate(
-                document_id=doc_uuid,
-                content=chunk.page_content,
-                embedding_status=DocumentProcessingStatus.PENDING,
-                num_tokens=len(chunk.page_content.split()),  # Simple token count
+            # Update document status to PROCESSING
+            await document.update(
+                session,
+                db_obj=db_document,
+                obj_in={
+                    "processing_status": DocumentProcessingStatus.PROCESSING,
+                    "error_message": None,
+                },
             )
-            db_chunk = await document_chunk.create(db, obj_in=chunk_data)
-            await db.flush()
-            await db.refresh(db_chunk)
-            logger.info(f"Created chunk {db_chunk.id} for document {document_id}")
+            await session.commit()
+            await session.refresh(db_document)
 
-            # Enqueue embedding task for each chunk
-            embed_chunk.delay(str(db_chunk.id), db_chunk.content)
+            # Load document
+            loader = None
+            file_extension = Path(file_path).suffix.lower()
+            if file_extension == ".pdf":
+                loader = PyPDFLoader(file_path)
+            elif file_extension == ".txt":
+                loader = TextLoader(file_path)
+            else:
+                raise ValueError(f"Unsupported file type: {file_extension}")
 
-        # Update document status to COMPLETED if all chunks are enqueued
-        await document.update(
-            db,
-            db_obj=db_document,
-            obj_in={
-                "processing_status": DocumentProcessingStatus.COMPLETED,
-            },
-        )
-        await db.commit()
-        logger.info(f"Document {document_id} processing completed and chunks enqueued.")
+            langchain_documents = loader.load()
+
+            # Split document into chunks
+            text_splitter = RecursiveCharacterTextSplitter(
+                chunk_size=1000, chunk_overlap=200
+            )
+            chunks = text_splitter.split_documents(langchain_documents)
+
+            # Create document chunks in DB and enqueue embedding tasks
+            for i, chunk in enumerate(chunks):
+                chunk_data = DocumentChunkCreate(
+                    document_id=doc_uuid,
+                    content=chunk.page_content,
+                    embedding_status=DocumentProcessingStatus.PENDING,
+                    num_tokens=len(chunk.page_content.split()),  # Simple token count
+                )
+                db_chunk = await document_chunk.create(session, obj_in=chunk_data)
+                await session.flush()
+                await session.refresh(db_chunk)
+                logger.info(f"Created chunk {db_chunk.id} for document {document_id}")
+
+                # Enqueue embedding task for each chunk
+                embed_chunk.delay(str(db_chunk.id), db_chunk.content)
+
+            # Update document status to COMPLETED if all chunks are enqueued
+            await document.update(
+                session,
+                db_obj=db_document,
+                obj_in={
+                    "processing_status": DocumentProcessingStatus.COMPLETED,
+                },
+            )
+            await session.commit()
+            logger.info(f"Document {document_id} processing completed and chunks enqueued.")
 
     except Exception as e:
         logger.error(f"Error processing document {document_id}: {e}", exc_info=True)
@@ -116,49 +109,42 @@ async def process_document(self, document_id: str, file_path: str):
         if Path(file_path).exists():
             os.remove(file_path)
             logger.info(f"Cleaned up temporary file: {file_path}")
-        if db:
-            await db.close()
 
 
 @celery_app.task(name="embed_chunk_task", bind=True)
 async def embed_chunk(self, chunk_id: str, chunk_content: str):
     db: AsyncSession = None
     try:
-        async for session in get_db_session():
-            db = session
-            break
+        db = async_session_maker()
+        async with db as session:
+            chunk_uuid = UUID(chunk_id)
+            db_chunk = await document_chunk.get(session, id=chunk_uuid)
 
-        if not db:
-            raise Exception("Could not get database session")
+            if not db_chunk:
+                logger.error(f"Chunk with ID {chunk_id} not found.")
+                return
 
-        chunk_uuid = UUID(chunk_id)
-        db_chunk = await document_chunk.get(db, id=chunk_uuid)
+            # Initialize embedding model and ChromaDB
+            embeddings = OpenAIEmbeddings(api_key=settings.OPENAI_API_KEY, model="text-embedding-ada-002")
+            # Assuming ChromaDB is initialized to a persistent path
+            # For simplicity, we'll re-initialize it here. In a real app, manage this more robustly.
+            # Ensure CHROMA_DB_PATH is configured in settings
+            chroma_client = Chroma(persist_directory=settings.CHROMA_DB_PATH, embedding_function=embeddings)
 
-        if not db_chunk:
-            logger.error(f"Chunk with ID {chunk_id} not found.")
-            return
+            # Add chunk to ChromaDB
+            # ChromaDB expects a list of documents, so we wrap the content
+            chroma_client.add_texts(texts=[chunk_content], metadatas=[{"chunk_id": str(chunk_uuid), "document_id": str(db_chunk.document_id)}])
 
-        # Initialize embedding model and ChromaDB
-        embeddings = OpenAIEmbeddings(api_key=settings.OPENAI_API_KEY, model="text-embedding-ada-002")
-        # Assuming ChromaDB is initialized to a persistent path
-        # For simplicity, we'll re-initialize it here. In a real app, manage this more robustly.
-        # Ensure CHROMA_DB_PATH is configured in settings
-        chroma_client = Chroma(persist_directory=settings.CHROMA_DB_PATH, embedding_function=embeddings)
-
-        # Add chunk to ChromaDB
-        # ChromaDB expects a list of documents, so we wrap the content
-        chroma_client.add_texts(texts=[chunk_content], metadatas=[{"chunk_id": str(chunk_uuid), "document_id": str(db_chunk.document_id)}])
-
-        # Update chunk status to COMPLETED
-        await document_chunk.update(
-            db,
-            db_obj=db_chunk,
-            obj_in={
-                "embedding_status": DocumentProcessingStatus.COMPLETED,
-            },
-        )
-        await db.commit()
-        logger.info(f"Chunk {chunk_id} embedded and stored in ChromaDB.")
+            # Update chunk status to COMPLETED
+            await document_chunk.update(
+                session,
+                db_obj=db_chunk,
+                obj_in={
+                    "embedding_status": DocumentProcessingStatus.COMPLETED,
+                },
+            )
+            await session.commit()
+            logger.info(f"Chunk {chunk_id} embedded and stored in ChromaDB.")
 
     except Exception as e:
         logger.error(f"Error embedding chunk {chunk_id}: {e}", exc_info=True)
@@ -171,6 +157,3 @@ async def embed_chunk(self, chunk_id: str, chunk_content: str):
                 },
             )
             await db.commit()
-    finally:
-        if db:
-            await db.close()
